Add a parameter to specify no output timeout
This allows users to specify a higher `no_output_timeout` for the git
push to Heroku, which could be a slow task.

If any part of the build on Heroku takes more than 10 minutes and
produces no output, CircleCI will consider this job failed. Heroku will
still continue to build.